end game

+++ Assets/Scripts/Hittable.cs
@@ -1,5 +1,5 @@
 public interface Hittable
 {
-    // Returns experience gained if killed, -1 otherwise
-    public int Hit(float damages);
+    // Returns damages done
+    public float Hit(float damages);
 }
